backend/requirements.txt
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
